fix(timetable): centralize course schedule rules
